# front_website/src/app/middleware/api_service/model/user.ts
/**
 * ICC PROJECT
 * Welcome to use ICC product
 *
 * OpenAPI spec version: 1.0.0
 * Contact: 1@12.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface User { 
    id?: number;
    username?: string;
    firstName?: string;
    lastName?: string;
    email?: string;
    password?: string;
    phone?: string;
    /**
     * User Status
     */
    userStatus?: number;
}