# go_server_main/src/go_server/swagger/model_user.go
/*
 * ICC PROJECT
 *
 * Welcome to use ICC product
 *
 * API version: 1.0.0
 * Contact: 
 * Generated by: lcu
 */

package swagger

type User struct {

	Id int64 `json:"id,omitempty"`

	Username string `json:"username,omitempty"`

	FirstName string `json:"firstName,omitempty"`

	LastName string `json:"lastName,omitempty"`

	Email string `json:"email,omitempty"`

	Password string `json:"password,omitempty"`

	Phone string `json:"phone,omitempty"`

	// User Status
	UserStatus int32 `json:"userStatus,omitempty"`
}